requirements.txt
openai-whisper
torch
ffmpeg-python
setuptools-rust
scipy
numpy
ty
ruff